# backend/Fms/openapi/index.yaml
openapi: 3.0.1
info:
  title: FMS
  version: v1
servers:
  - url: http://localhost:3333
paths:
...
  '/api/transactions/{id}':
    get:
      tags:
        - Transaction
      summary: Get transaction
      operationId: getTransaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicClientError'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
    post:
      tags:
        - Transaction
      summary: Update transaction
      operationId: updateTransaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionUpsertRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/TransactionUpsertRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/TransactionUpsertRequest'
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicClientError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
    delete:
      tags:
        - Transaction
      summary: Delete transaction
      operationId: deleteTransaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
...
  /api/transactions/categories:
    put:
      tags:
        - TransactionCategory
      summary: Create user transaction category
      operationId: createUserTransactionCategory
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionCategoryUpsertRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/TransactionCategoryUpsertRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/TransactionCategoryUpsertRequest'
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: integer
                format: int32
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicClientError'
    get:
      tags:
        - TransactionCategory
      summary: Get list of current user's transaction categories
      operationId: listUserTransactionCategories
      parameters:
        - name: Offset
          in: query
          schema:
            type: integer
            format: int32
        - name: Limit
          in: query
          schema:
            type: integer
            format: int32
        - name: IncludeOwner
          in: query
          schema:
            type: boolean
            example: true
          example: true
        - name: Query
          in: query
          schema:
            type: string
            example: mple catego
          example: mple catego
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionCategoryListResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicClientError'
  '/api/organizations/{organizationId}/transactions/categories':
    put:
      tags:
        - TransactionCategory
      summary: Create organization transaction category
      operationId: createOrganizationTransactionCategory
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionCategoryUpsertRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/TransactionCategoryUpsertRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/TransactionCategoryUpsertRequest'
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: integer
                format: int32
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicClientError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
    get:
      tags:
        - TransactionCategory
      summary: Get list of organization's transaction categories
      operationId: listOrganizationTransactionCategories
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: Offset
          in: query
          schema:
            type: integer
            format: int32
        - name: Limit
          in: query
          schema:
            type: integer
            format: int32
        - name: IncludeOwner
          in: query
          schema:
            type: boolean
            example: true
          example: true
        - name: Query
          in: query
          schema:
            type: string
            example: mple catego
          example: mple catego
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionCategoryListResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicClientError'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
  '/api/workspaces/{workspaceId}/transactions/categories':
    put:
      tags:
        - TransactionCategory
      summary: Create workspace transaction category
      operationId: createWorkspaceTransactionCategory
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionCategoryUpsertRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/TransactionCategoryUpsertRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/TransactionCategoryUpsertRequest'
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: integer
                format: int32
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicClientError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
    get:
      tags:
        - TransactionCategory
      summary: Get list of workspace transaction categories
      operationId: listWorkspaceTransactionCategories
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: Offset
          in: query
          schema:
            type: integer
            format: int32
        - name: Limit
          in: query
          schema:
            type: integer
            format: int32
        - name: IncludeOwner
          in: query
          schema:
            type: boolean
            example: true
          example: true
        - name: Query
          in: query
          schema:
            type: string
            example: mple catego
          example: mple catego
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionCategoryListResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicClientError'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
  '/api/transactions/categories/{id}':
    get:
      tags:
        - TransactionCategory
      summary: Get transaction category
      operationId: getTransactionCategory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionCategoryResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
    post:
      tags:
        - TransactionCategory
      summary: Update transaction category
      operationId: updateTransactionCategory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionCategoryUpsertRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/TransactionCategoryUpsertRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/TransactionCategoryUpsertRequest'
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicClientError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
    delete:
      tags:
        - TransactionCategory
      summary: Delete transaction category
      operationId: deleteTransactionCategory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '500':
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '200':
          description: Success
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicError'
...
components:
  schemas:
...
    AccountResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserResponse'
        organization:
          $ref: '#/components/schemas/OrganizationShortResponse'
      additionalProperties: false
...
    OrganizationShortResponse:
      required:
        - id
        - name
      type: object
      properties:
        id:
          type: integer
          format: int32
          example: 243
        name:
          minLength: 1
          type: string
          example: Cool Organization Inc.
      additionalProperties: false
...
    PublicClientError:
      required:
        - description
      type: object
      properties:
        description:
          minLength: 1
          type: string
          example: Something bad happened
        validationErrors:
          type: array
          items:
            type: string
          nullable: true
          example:
            - Email can't contain less than 3 characters
            - Password can't be empty
      additionalProperties: false
    PublicError:
      required:
        - description
      type: object
      properties:
        description:
          minLength: 1
          type: string
          example: Something bad happened
      additionalProperties: false
...
    TransactionCategoryKind:
      enum:
        - INCOME
        - EXPENSE
        - MIXED
      type: string
    TransactionCategoryListResponse:
      required:
        - items
        - totalCount
      type: object
      properties:
        totalCount:
          type: integer
          format: int32
          example: 165
        items:
          type: array
          items:
            $ref: '#/components/schemas/TransactionCategoryResponse'
      additionalProperties: false
    TransactionCategoryResponse:
      required:
        - id
        - kind
        - name
        - ownerAccount
      type: object
      properties:
        id:
          type: integer
          format: int32
          example: 5359
        name:
          minLength: 1
          type: string
          example: Salary
        ownerAccount:
          $ref: '#/components/schemas/AccountResponse'
        kind:
          $ref: '#/components/schemas/TransactionCategoryKind'
        uiColor:
          type: string
          nullable: true
          example: FF0022
        workspace:
          $ref: '#/components/schemas/WorkspaceShortResponse'
      additionalProperties: false
    TransactionCategoryUpsertRequest:
      required:
        - kind
        - name
      type: object
      properties:
        name:
          maxLength: 255
          minLength: 3
          type: string
          example: Food
        kind:
          $ref: '#/components/schemas/TransactionCategoryKind'
        uiColor:
          type: string
          nullable: true
          example: FF7E38
      additionalProperties: false
...
    TransactionResponse:
      required:
        - amount
        - category
        - creationTimestamp
        - creationUser
        - id
        - timestamp
        - workspace
      type: object
      properties:
        id:
          type: integer
          format: int32
        workspace:
          $ref: '#/components/schemas/WorkspaceShortResponse'
        category:
          $ref: '#/components/schemas/TransactionCategoryResponse'
        amount:
          type: integer
          format: int32
          example: -530
        description:
          type: string
          nullable: true
          example: Some optional description
        timestamp:
          type: string
          format: date-time
        creationTimestamp:
          type: string
          format: date-time
        creationUser:
          $ref: '#/components/schemas/UserResponse'
        user:
          $ref: '#/components/schemas/UserResponse'
      additionalProperties: false
...
    TransactionUpsertRequest:
      required:
        - amount
        - categoryId
      type: object
      properties:
        categoryId:
          type: integer
          format: int32
          example: 6402
        amount:
          type: integer
          format: int32
          example: 640
        description:
          maxLength: 2048
          type: string
          nullable: true
          example: Some different optional description
        timestamp:
          type: string
          format: date-time
          nullable: true
          example: 
        userId:
          type: integer
          format: int32
          nullable: true
          example: 
      additionalProperties: false
...
    UserResponse:
      required:
        - firstName
        - id
        - lastName
      type: object
      properties:
        id:
          type: integer
          format: int32
          example: 25
        firstName:
          minLength: 1
          type: string
          example: John
        lastName:
          minLength: 1
          type: string
          example: Snow
      additionalProperties: false
...
    WorkspaceKind:
      enum:
        - PRIVATE
        - SHARED
      type: string
...
    WorkspaceShortResponse:
      required:
        - id
        - kind
        - name
      type: object
      properties:
        id:
          type: integer
          format: int32
          example: 1402
        name:
          minLength: 1
          type: string
          example: Family Workspace
        kind:
          $ref: '#/components/schemas/WorkspaceKind'
      additionalProperties: false
...
  securitySchemes:
    Bearer:
      type: apiKey
      description: "JWT Authorization header using the Bearer scheme.\n\nEnter 'Bearer' [space] and then your token in the text input below.\nExample: 'Bearer 12345abcdef'"
      name: Authorization
      in: header
security:
  - Bearer: [ ]